Command Pattern implementieren damit Züge rückgängig gemacht werden (#8)
* Basic structure for board creation.
- loads board from default csv which sets up board.
There is a UI side and internal storage side.
The game will send out moves to the chess engine once its running.
It is fully capable of understanding if a move is valid, but the server needs to verify if the move is correct nonetheless

* Added logging to program, adjusted pom.xml to be compilable with mvn aswell.
Minor changes to png loading logic (fallback placeholder)

* Added Constants class which contains paths for loading images and configurations
Revamped a lot of documentation
Revamped PngLoader to be a singleton that has a key-value map storing a name of a image and the reference for usage

* Current state

* Finished movement logic and making simple moves.
Added a simple sound engine with SoundManager + Sound + SoundType.
This will be expanded later

* Fixed possible movement generation bugs, adjusted chessboard to update the corresponding ChessBoardTilePane when moving a piece.
Cleaned up some code in ChessPiece and Position to fix a bug with isOccupied and adjusted the BoardController accordingly

* Restructured the the project to have 2 seperate projects in 1 repo. there is a parent pom.xml and for each component a seperate jar.
The jars will be created via the parent pom.xml in the CI pipeline later on

* re added the client to the branch

* Had to fix some project setup related issues otherwise the project wont start

* More project cleanup

* fixing the project (client side

* Overhaul of game logic. Added command pattern, adjusted board controller.
gamemaster controls flow of the game and interfaces with  controller now.
movement can determine if it can move or capture. Small bug fixes aswell

* Added new Commans and kept working on the movement behaviour.

Networking implementation started.
Promotion / Casteling + other moves possible, mouse input simplified

* Added new sounds + finished commands
2 players can locally play the game until one wins. checkmate / check is still buggy
after win menu exists now, can save history, rematch or exit to main menu. Adjusted code some more fixed some nullpointer exceptions and adjusted the way strategies work in the chesssim

* Code fixup, added documentation
networking code updated on server side, rooms can be created and theoretically moves are sent to the player
client side is not ready to recieve moves / listen properly causing an issue
adjusted MainController.java
added NetworkPlayer
Some extra popups were created
networking code is all over the place

* Added documentation for the server classes
Adjusted the gameManger so that moves can be undone
Removed unneeded imports / fixed typos
General cleanup

---------

Co-authored-by: Rene Polak <ic24b022@technikum-wien.at>